Added basic shop interface

+++ Assets/Scripts/InventoryItem.cs
@@ -1,13 +1,11 @@
-using UnityEngine;
-
 public struct InventoryItem
 {
     public string id;
     public float scaleFactor;
     public int materialIndex;
-    public float price;
+    public int price;
 
-    public InventoryItem(string id, float scaleFactor, int materialIndex, float price)
+    public InventoryItem(string id, float scaleFactor, int materialIndex, int price)
     {
         this.id = id;
         this.scaleFactor = scaleFactor;
